--- a/src/biomcp/http_client.py
+++ b/src/biomcp/http_client.py
@@ -1,5 +1,4 @@
 import csv
-import hashlib
 import json
 import os
 import ssl
@@ -52,14 +51,21 @@ def get_cache() -> Cache:
     return _cache
 
 
-# noinspection PyTypeChecker
 def generate_cache_key(method: str, url: str, params: dict) -> str:
-    sha256_hash = hashlib.sha256()
-    params_dump: str = json.dumps(params, sort_keys=True)
-    key_source: str = f"{method.upper()}:{url}:{params_dump}"
-    data: bytes = key_source.encode("utf-8")
-    sha256_hash.update(data)
-    return sha256_hash.hexdigest()
+    """Generate cache key using Python's built-in hash function for speed."""
+    # Handle simple cases without params
+    if not params:
+        return f"{method.upper()}:{url}"
+
+    # Use Python's built-in hash with a fixed seed for consistency
+    # This is much faster than SHA256 for cache keys
+    params_str = json.dumps(params, sort_keys=True, separators=(",", ":"))
+    key_source = f"{method.upper()}:{url}:{params_str}"
+
+    # Use Python's hash function with a fixed seed for deterministic results
+    # Convert to positive hex string for compatibility
+    hash_value = hash(key_source)
+    return f"{hash_value & 0xFFFFFFFFFFFFFFFF:016x}"
 
 
 def cache_response(cache_key: str, content: str, ttl: int):