--- a/src/biomcp/individual_tools.py
+++ b/src/biomcp/individual_tools.py
@@ -105,6 +105,7 @@ async def article_searcher(
         keywords=keywords,
         variants=variants,
         include_preprints=include_preprints,
+        include_cbioportal=include_cbioportal,
     )
 
     # Add cBioPortal summary if searching by gene