… CHANGELOG.md
… README.md
… docs/performance-optimizations.md
… mkdocs.yml
… src/biomcp/connection_pool.py
… src/biomcp/http_client_simple.py
… src/biomcp/metrics.py
… src/biomcp/prefetch.py
… src/biomcp/request_batcher.py
… src/biomcp/shared_context.py
… src/biomcp/utils/request_cache.py
… src/biomcp/variants/cbioportal_search.py
… test_performance_analysis.md
… tests/tdd/test_connection_pool.py
… tests/tdd/test_metrics.py
… tests/tdd/test_request_batcher.py.disabled
… tests/tdd/test_retry.py
… tests/tdd/test_shared_context.py.disabled
… uv.lock