Major performance improvements achieving ~3x faster test execution (120s → 42s):

      ## Connection Management
      - Implement event-loop aware connection pooling with weak references
      - Add proper lifecycle management to prevent event loop issues
      - Reduce connection overhead by reusing HTTP clients across requests
      - Fix async context manager for FastMCP lifespan

      ## API Call Optimization
      - Parallelize autocomplete calls in article searches (3x speedup)
      - Fix sequential cBioPortal fetching with proper batching
      - Implement request batching for multiple API calls
      - Add shared context to eliminate redundant gene validations
      - Optimize bioRxiv date range from 30 to 365 days per user feedback

      ## Caching & Performance
      - Add LRU caching with configurable size limits
      - Implement prefetching for common genes, diseases, and chemicals
      - Add hash function optimization for cache keys
      - Introduce conditional metrics to reduce overhead

      ## Test Execution
      - Enable parallel test execution with pytest-xdist
      - Add pagination support for Europe PMC searches
      - Optimize test fixtures and reduce redundant API calls

      ## Code Quality
      - Extract performance constants to centralized configuration
      - Add comprehensive documentation for all optimizations
      - Create unit tests for connection pooling and shared context
      - Fix all linting and type checking issues

      ## Performance Metrics
      - Test suite execution: 120s → 42s (65% improvement)
      - API call reduction: ~40% fewer calls with shared context
      - Memory usage: Optimized with weak references and size-limited caches
      - Connection reuse: 20 keepalive connections per pool

      Breaking Changes: None - all optimizations maintain backward compatibility